Partially integrated new inventory and UI
Partially integrated inventory into UI, still need to re-do how
tooltips work and fix some quirks that appeared. Have to rename some
slots int he character equipment and there might be some logic issues
with how I reset the menu order if the tooltip has to be always last in
the order. Perhaps I can add an on-enable to it to ensure it’s last.